test(api): update api tests for api change
